@@ -12,10 +12,11 @@
     "express": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
     "middleware": "^1.0.0",
-    "react": "^19.1.1",
+    "react": "^19.2.0",
     "react-dom": "^19.1.1",
     "react-router-dom": "^6.30.1",
     "react-scripts": "5.0.1",
+    "scripts": "^0.1.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
